balloon-hash: bump `crypto-bigint` to v0.5; MSRV 1.65 (#381)

+++ argon2/src/params.rs
@@ -174,7 +174,7 @@ impl Params {
 
     /// Get the number of blocks required given the configured `m_cost` and `p_cost`.
     pub fn block_count(&self) -> usize {
-        (self.segment_length() * self.lanes() * SYNC_POINTS) as usize
+        self.segment_length() * self.lanes() * SYNC_POINTS
     }
 }
